+++ lightningd/subd.c
@@ -17,7 +17,6 @@
 #include <lightningd/subd.h>
 #include <sys/socket.h>
 #include <sys/wait.h>
-#include <wire/common_wiregen.h>
 #include <wire/wire_io.h>
 
 /* Carefully move fd *@from to @to: on success *from set to to */
@@ -816,8 +815,7 @@ void subd_send_msg(struct subd *sd, const u8 *msg_out)
 	u16 type = fromwire_peektype(msg_out);
 	/* FIXME: We should use unique upper bits for each daemon, then
 	 * have generate-wire.py add them, just assert here. */
-	assert(!strstarts(common_wire_name(type), "INVALID") ||
-	       !strstarts(sd->msgname(type), "INVALID"));
+	assert(!strstarts(sd->msgname(type), "INVALID"));
 	msg_enqueue(sd->outq, msg_out);
 }
